requirements.txt
pypdf
langchain
torch
accelerate
bitsandbytes
ctransformers
sentence-transformers
faiss_cpu
chainlit
huggingface_hub
langchain_community
streamlit
pillow
transformers
requests
googletrans==4.0.0-rc1
gTTS
google-generativeai
python-docx
pytesseract
pdf2image
pymupdf
langchain-huggingface